+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Changes are grouped as follows
 
 ## [0.33.0] - 2020-12-02
 ### Added
-- Method `get_data` on the `FunctionSchedule`-class. 
-- Method `get_data` on the `FunctionScheduleAPI`-class.
+- Method `get_input_data` on the `FunctionSchedule`-class. 
+- Method `get_input_data` on the `FunctionScheduleAPI`-class.
 
 ## [0.32.0] - 2020-12-03
 ### Fixed